Add basic typing animation for phase changes
I couldn't think of anything better for this one, other than the
rediculous skuomorphism of rotating the text in 3D like a roll display
(which I also couldn't work out how to do in canvas) so I went with
a tried and tested typewriter effect.

I'd like for this to have a cursor and occasionally make a typo, but for
now it'll do great.